Update Example for HeaderMenu, fix Autocomplete export
